requirements.txt

aniso8601==1.2.0
asn1crypto==0.23.0
backports.shutil-get-terminal-size==1.0.0
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
Click==7.0
cryptography==2.1.3
enum34==1.1.6
Flask==1.1.1
Flask-Ask==0.9.7
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.8
Flask-HTTPAuth==3.3.0
Flask-SocketIO==4.2.1
gunicorn==19.9.0
idna==2.6
ipaddress==1.0.18
itsdangerous==1.1.0
Jinja2==2.10.1
keyring==10.1
keyrings.alt==1.3
MarkupSafe==1.1.1
numpy==1.12.1
paho-mqtt==1.3.1
picamera==1.13
prompt-toolkit==1.0.16
pyasn1==0.1.9
pycparser==2.18
pycrypto==2.6.1
pycurl==7.43.0
pygobject==3.22.0
pyOpenSSL==17.0.0
python-dateutil==2.6.1
python-engineio==3.9.3
python-socketio==4.3.1
pyxdg==0.25
PyYAML==3.12
requests==2.11.1
RPi.GPIO==0.6.5
SecretStorage==2.3.1
six==1.12.0
urllib3==1.22
wcwidth==0.1.7
Werkzeug==0.15.5
